requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
flask-bootstrap5==0.1.dev1
Flask-WTF==1.2.2
fonttools==4.57.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
